# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Access Control API
  description: A comprehensive API for managing users, roles, and privileges in a secure, production-ready Spring Boot application.
  version: "1.0.0"
servers:
  - url: http://localhost:8080/api
    description: Development server
paths:
...
  /admin/users:
    get:
      summary: Get Users
      description: Retrieves a paginated list of users, with support for filtering and sorting. Requires USER_MANAGEMENT_READ privilege.
      tags: [Admin - User Management]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 0 }
        - in: query
          name: size
          schema: { type: integer, default: 10 }
        - in: query
          name: sort
          schema: { type: array, items: { type: string } }
          example: ["username,asc"]
        - in: query
          name: search
          schema: { type: string }
      responses:
        '200':
          description: Users fetched successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JsonResponsePagedUsers'
              examples:
                Success:
                  value:
                    status: 200
                    message: "Users fetched successfully"
                    data:
                      content:
                        - textId: "user-text-id-123"
                          username: "admin"
                          email: "admin@meta.com"
                          enabled: true
                          roles: ["ROLE_ADMIN"]
                          createdAt: 1672531200000
                          createdBy: "system"
                      pageNumber: 0
                      pageSize: 10
                      totalElements: 1
                      totalPages: 1
                      isLast: true
        '403':
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JsonResponseError'
    post:
      summary: Create User
      description: Creates a new user with a default password. Requires USER_MANAGEMENT_WRITE privilege.
      tags: [Admin - User Management]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JsonResponseUser'
        '409':
          description: Username or email already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JsonResponseError'
  /admin/users/{textId}:
    get:
      summary: Get User Details
      description: Retrieves the full details for a specific user. Requires USER_MANAGEMENT_READ privilege.
      tags: [Admin - User Management]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: textId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: User details fetched successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JsonResponseUser'
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JsonResponseError'
    put:
      summary: Update User
      description: Updates a user's roles and/or enabled status. Requires USER_MANAGEMENT_WRITE privilege.
      tags: [Admin - User Management]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: textId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JsonResponseUser'
    delete:
      summary: Delete User
      description: Deletes a user. Requires USER_MANAGEMENT_WRITE privilege.
      tags: [Admin - User Management]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: textId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: User deleted successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JsonResponse'
        '400':
          description: Cannot delete your own account.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JsonResponseError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateUserRequest:
      type: object
      properties: { username: { type: string }, email: { type: string }, roleTextIds: { type: array, items: { type: string } } }
    UpdateUserRequest:
      type: object
      properties: { roleTextIds: { type: array, items: { type: string } }, enabled: { type: boolean } }
...
    Gender:
      type: string
      enum: [MALE, FEMALE, PREFER_NOT_TO_SAY, OTHER]

    # Response Bodies & Payloads
    JsonResponse:
      type: object
      properties: { status: { type: integer }, message: { type: string }, data: { type: object, nullable: true } }
    JsonResponseError:
      type: object
      properties: { status: { type: integer }, message: { type: string }, data: { type: object, nullable: true } }
...
    UserResponse:
      type: object
      properties:
        textId: { type: string }
        username: { type: string }
        email: { type: string }
        enabled: { type: boolean }
        roles: { type: array, items: { type: string } }
        privileges: { type: array, items: { type: string }, nullable: true }
        profile: { $ref: '#/components/schemas/UserProfileResponse', nullable: true }
        createdAt: { type: integer, format: int64, nullable: true }
        createdBy: { type: string, nullable: true }
        updatedAt: { type: integer, format: int64, nullable: true }
        updatedBy: { type: string, nullable: true }
    UserProfileResponse:
      type: object
      properties:
        firstName: { type: string, nullable: true }
        lastName: { type: string, nullable: true }
        dateOfBirth: { type: string, format: date, nullable: true }
        gender: { $ref: '#/components/schemas/Gender', nullable: true }
        phoneNumber: { type: string, nullable: true }
        address: { type: string, nullable: true }
        avatarUrl: { type: string, nullable: true }
        timeZone: { type: string, nullable: true }
        locale: { type: string, nullable: true }
    PaginationResponse:
      type: object
      properties: { pageNumber: { type: integer }, pageSize: { type: integer }, totalElements: { type: integer }, totalPages: { type: integer }, isLast: { type: boolean } }
...
    JsonResponseUser:
      allOf: [ { $ref: '#/components/schemas/JsonResponse' }, { type: object, properties: { data: { $ref: '#/components/schemas/UserResponse' } } } ]
    JsonResponsePagedUsers:
      allOf: [ { $ref: '#/components/schemas/JsonResponse' }, { type: object, properties: { data: { allOf: [ { $ref: '#/components/schemas/PaginationResponse' }, { type: object, properties: { content: { type: array, items: { $ref: '#/components/schemas/UserResponse' } } } } ] } } } ]
...